Tweak foldexpr to be less surprising
This removes the need for the hack that caused folds to close when in insert
mode and should contain less surprising behavior when working with folds.

We special case an indent of 0 because previously it would be no fold at all,
which suddenly became a closed fold when you insert text underneath it.

In addition, change the behavior on blank lines to be equal to the previous
line rather than undefined, which closer matches what is intended (blank lines
should be part of whatever is above them).